pyproject.toml
[project]
name = "classifier"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "joblib>=1.5.2",
    "lightgbm>=4.6.0",
    "numpy>=2.3.5",
    "openai>=2.9.0",
    "optuna>=4.6.0",
    "pandas>=2.3.3",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "rich>=14.2.0",
    "scikit-learn>=1.7.2",
    "tabulate>=0.9.0",
    "xgboost>=3.1.2",
    "altair>=5.0.0",
]
